coding-test/q5/requirements.txt
# Core Framework
streamlit

# OCR - Gemini Vision API
google-generativeai
Pillow

# LangChain and LLM
langchain
langchain-core
langchain-google-genai
langgraph
pydantic

# Embeddings and Vector Search
sentence-transformers
numpy

# Database and Storage
python-dotenv
